chore: move aum to own module
